move: User - tests
