+++ app/tests/users/test_employee.py
@@ -1,7 +1,7 @@
 from typing import Type
 
 from entities.users.models_user import EmployeeModel
-from tests.abstract_user import UserAbstract
+from tests.users.abstract_user import UserAbstract
 from entities.users.schemas_users import EmployeeInSchema, EmployeeInOptionalSchema
 from services.stub_init_service import LAST_NAMES
